+++ typescript/packages/plugins/solana-magiceden/src/index.ts
@@ -1,1 +1,2 @@
-export * from "./plugin";
+export * from "./magic-eden.plugin";
+export * from "./parameters";
