+++ typescript/packages/plugins/solana-magiceden/src/magic-eden.plugin.ts
@@ -1,0 +1,10 @@
+import { Chain, PluginBase } from "@goat-sdk/core";
+import { MagicEdenService } from "./magic-eden.service";
+
+export class MagicEdenPlugin extends PluginBase {
+    constructor(apiKey?: string) {
+        super("magic-eden", [new MagicEdenService(apiKey)]);
+    }
+
+    supportsChain = (chain: Chain) => chain.type === "solana";
+}
